Refactored JSON drivers location
