+++ autodict/json_drivers/__init__.py
@@ -1,0 +1,5 @@
+"""JSON Drivers to serialize/deserialize AutoDict objects
+"""
+
+from autodict.json_drivers.base import (JSONAutoDict, JSONDriver,
+                                        DefaultJSONDriver)
